Switch to nested if
A 200 line method isn't great, but the previous error came from tring to
user return at the wrong point. Someone could miss the `next` call and
add to the bottom of the each loop (between the if/end and the loop end)
without realising it won't always be called.

This isn't great but at least it is consistent.